Amend proper types on form components
