+++ frontend-voucher/src/constants/globalTypes.ts
@@ -1,0 +1,11 @@
+// Place any reuseable types in this file
+import { Dayjs } from 'dayjs';
+
+export type voucherFormValues = {
+  action: string;
+  category: string;
+  description: string;
+  minSpending: number;
+  startDate: Dayjs;
+  expiryDate: Dayjs;
+};
